Renaming macro file.
